achievements for ios

+++ game-services/core/src/com/submarine/gameservices/GameServices.java
@@ -13,7 +13,7 @@ public interface GameServices {
 
     void unlockAchievement(String achievementId);
 
-    void incrementAchievement(String achievementId, int incrementAmount);
+    void incrementAchievement(String achievementId, double incrementAmount);
 
     void showAchievements();
